backend/requirements.txt
Django==5.2.13
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg[binary]==3.2.13
python-decouple==3.8
gunicorn==23.0.0
